requirements.txt
fastapi==0.111.1
uvicorn==0.23.2
requests==2.31.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
pydantic>=2.0.0
sqlalchemy>=2.0.0
osmnx>=1.9.0
networkx>=3.0
pillow>=10.0
matplotlib>=3.8.0
google-genai
